Consolidated `help` and `help_with` into `help` impled on `Topic` (Help renamed)
