+++ src/gui/terminalgui.rs
@@ -1,9 +1,11 @@
 use std::io;
 use const_format::concatcp;
+use help::Topic;
 use thiserror::Error;
 use crate::{character::{Chr, Health}, conditions::Condition, saver::Saver, tracker::{self, Tracker}};
 
 mod parser;
+mod help;
 
 const CLEAR: &str  = "\x1B[2J\n";
 const SPACER: &str = "  ";
@@ -106,8 +108,7 @@ enum Command {
         health: Option<u32>
     },
     RmCond { character: String, cond: Condition },
-    Help,
-    HelpWith(Help),
+    Help(Topic),
 }
 
 fn execute_command<S: Saver>(t: &mut Tracker<S>, cmd: Command) -> tracker::Result<()> {
@@ -143,136 +144,10 @@ fn execute_command<S: Saver>(t: &mut Tracker<S>, cmd: Command) -> tracker::Resul
 
             Ok(())
         },
-        Command::Help => {
-            help();
-            Ok(())
-        }
-        Command::HelpWith(cmd) => {
-            help_with(cmd);
+        Command::Help(topic) => {
+            topic.help();
             Ok(())
         }
     }
 }
 
-const HELP_HEADER: &str = "Instructions";
-const ITEM: &str = " - ";
-const INDENT: &str = "   ";
-const HELP_HELP: &str = concatcp!(
-    ITEM, 
-    parser::command_strs::HELP, 
-    " [<command>]: explains a given command. If none is given, gives this overview."
-);
-const HELP_END_TURN: &str = concatcp!(
-    ITEM, 
-    parser::command_strs::END_TURN, 
-    ": ends the current turn"
-);
-const HELP_ADD: &str = concatcp!(
-    ITEM, 
-    parser::command_strs::ADD, 
-    ": adds a character.",
-);
-const HELP_REMOVE: &str = concatcp!(
-    ITEM,
-    parser::command_strs::REMOVE,
-    " <character>: removes a character from the tracker"
-);
-const HELP_MODIFY: &str = concatcp!(
-    ITEM,
-    parser::command_strs::MODIFY,
-    " <character> [options]: modifies one or more properties of a character.",
-);
-const HELP_CONDITION: &str = concatcp!(
-    ITEM,
-    parser::command_strs::CONDITION,
-    " <condition command>: do `help cond` for details.",
-);
-
-const HELP: &str = concatcp!(
-    HELP_HEADER, "\n", 
-    HELP_HELP, "\n",
-    HELP_END_TURN, "\n",
-    HELP_ADD, "\n",
-    HELP_REMOVE, "\n",
-    HELP_MODIFY, "\n",
-    HELP_CONDITION
-);
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum Help {
-    Help,
-    EndTurn,
-    Add,
-    Remove,
-    Modify,
-    Condition,
-}
-
-fn help() {
-    println!("{CLEAR}");
-    println!("{HELP}");
-
-    let mut buff = String::new();
-    let stdin = io::stdin();
-
-    let _ = stdin.read_line(&mut buff);
-}
-
-fn pause() {
-    let mut buff = String::new();
-    let stdin = io::stdin();
-
-    let _ = stdin.read_line(&mut buff);
-}
-
-
-const HELP_WITH_HELP: &str = concatcp!(
-    parser::command_strs::HELP, " [<command>]:\n\
-    \n\
-    Gives helpful information on commands. If no argument is given, a general\n\
-    command guide and summary list of commands is printed. `help` can also\n\
-    give a more detailed description of a command by providing the command as\n\
-    an argument.\n\
-    \n\
-    Example: ", parser::command_strs::HELP, " add"
-);
-
-const HELP_WITH_END_TURN: &str = concatcp!(
-    parser::command_strs::END_TURN, ":\n\
-    \n\
-    Ends the current turn.
-    "
-);
-
-const HELP_WITH_ADD: &str = concatcp!(
-    parser::command_strs::ADD, " <init> <name> [<options>]:\n\
-    \n\
-    Adds a new character to the tracker.\n\
-    \n\
-    Optional features like tracking health can be added by including the\n\
-    corresponding option. The options are:\n\
-     - health/h <max health>: adds health tracking\n\
-     - player/p: marks the character as a player character\n\
-     - enemy/e: marks the character as an enemy character\n\
-    \n\
-    Example: add 24 Sarah -player -health 20
-    "
-);
-
-fn help_with(cmd: Help) {
-    println!("{CLEAR}");
-    match cmd {
-        Help::Help => println!("{HELP_WITH_HELP}"),
-        Help::EndTurn => println!("{HELP_WITH_END_TURN}"),
-        Help::Add => println!("{HELP_WITH_ADD}"),
-        Help::Remove => todo!(),
-        Help::Modify => todo!(),
-        Help::Condition => todo!(),
-    };
-
-    println!();
-
-    println!("... press Enter to return");
-
-    pause();
-}
